Clock IT Free - Modern Nightstand Alarm
Transform your device into a functional and elegant nightstand alarm
clock! Better yet on your favourite dock.

Features of Clock IT include:
- Unlimited alarms featuring ten different alarm tones, personalised
labels & repeat alarm options
- Gently fade in alarms
- Gorgeous theme moderno
- Snooze feature with a countdown showing how much snooze time left
- Energy efficient - screen turns off when device is kept face down
- Alarm icon indicates alarm is ON/OFF
- Current location
- Current day of week
- 12 or 24-hour time format
- Set screen auto-lock on/off
- Designed for efficiency - low battery consumption
- Swipe to control brightness
- Automatic portrait & landscape modes
- HD Retina resolution
- 64 bit devices supported
- Free to download

__________________________________________
This app shows Ads on the clock when available.
__________________________________________

We're constantly improving Clock IT, and we value our customers'
opinions and feedback. If you want to send us your feedback or have a
problem with our software, please use the support feature within the
app or contact us directly at info@creoit.com before submitting a
comment to the App Store. Remember that we can solve most of the
problems, but we cannot respond to comments on the App Store.